+++ services/rds/struct_rule.go
@@ -19,8 +19,8 @@ package rds
 // Rule is a nested struct in rds response
 type Rule struct {
             Name     string `json:"Name" xml:"Name"`
-            RuleDescId     int `json:"RuleDescId" xml:"RuleDescId"`
+            RuleDescId     int64 `json:"RuleDescId" xml:"RuleDescId"`
             Title     string `json:"Title" xml:"Title"`
-            RuleId     int `json:"RuleId" xml:"RuleId"`
+            RuleId     int64 `json:"RuleId" xml:"RuleId"`
             Description     string `json:"Description" xml:"Description"`
 }
